core: Move text line helpers
Text-line helpers split byte streams, normalize line endings, and wrap
acquirable line sequences. They are neutral line primitives used by core
replacement payloads, batch matching, editor editing, and Git output
streaming.

Keeping them under utils made core.replacement import utilities after
mapped storage was moved into core. That left the core/utils cycle in place
for behavior that is not Git or filesystem integration.

This commit addresses that by moving utils.text to core.text_lines,
updating production callers, and removing the old utils module instead of
adding a shim.

Follow-up test commits will move the text-line tests and guard the core
boundary.
