+++ src/git_stage_batch/batch/absence_constraints.py
@@ -12,7 +12,7 @@ from ..exceptions import (
     MissingAnchorError as _MissingAnchorError,
 )
 from ..i18n import _
-from ..utils.text import normalize_line_endings
+from ..core.text_lines import normalize_line_endings
 from .merge_candidates import MergeResolution as _MergeResolution
 from .realized_boundaries import (
     boundary_choices_after_source_line as _boundary_choices_for_source_line,
