+++ src/git_stage_batch/batch/realized_file_content.py
@@ -10,7 +10,7 @@ from ..editor.line_endings import (
     detect_line_ending,
     restore_line_endings_in_chunks,
 )
-from ..utils.text import normalize_line_sequence_endings
+from ..core.text_lines import normalize_line_sequence_endings
 from .presence_constraints import satisfy_constraints
 from .realized_entry_storage import realized_entry_content_chunks
