+++ src/git_stage_batch/core/text_lines.py
